Create cs10.11.7.5:24pm
Write a single line of code to create the above multidimensional array of integers with the name pairs.  Do not put any of the data into the array just yet. Your answer should create the array only, not populate it with integers.....for...Consider the following table for each question.

  2	  4	  6	  8	  10
 12	 14 	 16	 18	  20
 22	 24	 26 	 28	  30
 32	 34	 36	 38	  40
where 2 is blue, 18 is green, and 34 is brown
To create a multidimensional array of integers with the name pairs without putting any data into the array, we can use the following single line of code:

C++
int pairs[4][5];
Use code with caution. Learn more
This code creates a two-dimensional array with 4 rows and 5 columns. The type of each element in the array is int.

Note that this code only creates the array, but does not populate it with any integers. To populate the array, we would need to use a loop to iterate over the elements and assign values to them.

Here is an example of how to populate the array with the values from the table:

C++
int pairs[4][5] = {
  {2, 4, 6, 8, 10},
  {12, 14, 16, 18, 20},
  {22, 24, 26, 28, 30},
  {32, 34, 36, 38, 40}
};
Use code with caution. Learn more
Once the array is populated, we can access its elements using the [] operator. For example, to access the element at row 2 and column 3, we would use the following code:

C++
int element = pairs[2][3];
Use code with caution. Learn more
This would assign the value 26 to the variable element.

Hope this helps! Let me know if you have any other questions.

profile picture
Write a single line of code to display the value in the array position indicated by the blue box in the above image. 
Note: Do not include an end line,,, where Consider the following table for each question.

  2	  4	  6	  8	  10
 12	 14 	 16	 18	  20
 22	 24	 26 	 28	  30
 32	 34	 36	 38	  40
where 2 is blue, 18 is green, and 34 is brown
To display the value in the array position indicated by the blue box in the above image, we can use the following single line of code:

C++
std::cout << pairs[1][0]; // 12
Use code with caution. Learn more
This code accesses the element at row 1 and column 0 of the array pairs and prints its value to the console.

Note that we do not need to include an end line in this case, as we are only printing a single value.

Here is an example of how to use the code:

C++
#include <iostream>

int main() {
  int pairs[4][5] = {
    {2, 4, 6, 8, 10},
    {12, 14, 16, 18, 20},
    {22, 24, 26, 28, 30},
    {32, 34, 36, 38, 40}
  };

  // Display the value in the array position indicated by the blue box.
  std::cout << pairs[1][0]; // 12

  return 0;
}
Use code with caution. Learn more
Output:

12
profile picture
Write a single line of code to change the value in the array position indicated by the red box in the above image to 100.  where...Consider the following table for each question.

  2	  4	  6	  8	  10
 12	 14 	 16	 18	  20
 22	 24	 26 	 28	  30
 32	 34	 36	 38	  40
where 2 is blue, 18 is red, and 34 is brown
To change the value in the array position indicated by the red box in the above image to 100, we can use the following single line of code:

C++
pairs[2][1] = 100;
Use code with caution. Learn more
This code accesses the element at row 2 and column 1 of the array pairs and assigns it the value 100.

Here is an example of how to use the code:

C++
#include <iostream>

int main() {
  int pairs[4][5] = {
    {2, 4, 6, 8, 10},
    {12, 14, 16, 18, 20},
    {22, 24, 26, 28, 30},
    {32, 34, 36, 38, 40}
  };

  // Change the value in the array position indicated by the red box to 100.
  pairs[2][1] = 100;

  // Display the value in the array position indicated by the red box.
  std::cout << pairs[2][1]; // 100

  return 0;
}
Use code with caution. Learn more
Output:

100
profile picture
Write a single line of code to create a variable named green and save the value in the array position indicated by the green box in the above image to that variable.  
Note: Use only one c++ statement to accomplish both tasks.
where ...Consider the following table for each question.

  2	  4	  6	  8	  10
 12	 14 	 16	 18	  20
 22	 24	 26 	 28	  30
 32	 34	 36	 38	  40
where 2 is blue, 18 is green, and 34 is red

1
Sure, here is a single line of code to create a variable named green and save the value in the array position indicated by the green box in the above image to that variable:

C++
int green = pairs[1][2];
Use code with caution. Learn more
This code declares a variable named green of type int and initializes it to the value of the element at row 1 and column 2 of the array pairs.

Here is an example of how to use the code:

C++
#include <iostream>

int main() {
  int pairs[4][5] = {
    {2, 4, 6, 8, 10},
    {12, 14, 16, 18, 20},
    {22, 24, 26, 28, 30},
    {32, 34, 36, 38, 40}
  };

  // Create a variable named green and save the value in the array position indicated by the green box to that variable.
  int green = pairs[1][2];

  // Display the value of the variable green.
  std::cout << green; // 16

  return 0;
}
Use code with caution. Learn more
Output:

16
Sources
www.chegg.com/homework-help/questions-and-answers/write-single-line-code-change-value-array-position-indicated-red-box-image-100--question-4-q117745775
profile picture
Two-dimensional arrays with initializer lists can omit (only) the length specification on the right.   where Consider the following table for each question.

  2	  4	  6	  8	  10
 12	 14 	 16	 18	  20
 22	 24	 26 	 28	  30
 32	 34	 36	 38	  40
where 2 is blue, 18 is green, and 34 is red
View other drafts























Yes, two-dimensional arrays with initializer lists can omit (only) the length specification on the right. This is because the compiler can infer the length of the rightmost dimension from the number of elements in the initializer list.

For example, the following code is valid:

C++
int array[][5] = {
  {1, 2, 3, 4, 5},
  {6, 7, 8, 9, 10},
  {11, 12, 13, 14, 15}
};
Use code with caution. Learn more
The compiler will infer that the length of the rightmost dimension is 5, because there are 5 elements in the initializer list for the first row of the array.

However, it is not possible to omit the length specification on the leftmost dimension:

C++
int array[][5] = {
  {1, 2, 3, 4},
  {6, 7, 8, 9}
};
Use code with caution. Learn more
This code is invalid, because the compiler cannot infer the length of the leftmost dimension from the initializer list.

It is generally considered to be good practice to explicitly specify the lengths of all dimensions of a multidimensional array, even if it is possible to omit some of them. This makes the code more readable and less error-prone.