Create 11.8.23..1800h
What does it mean to overload an operator in c++
View other drafts
























Operator overloading in C++ is a feature that allows you to give a new meaning to existing operators. This means that you can use the same operator to perform different operations on different data types. For example, you can overload the + operator to add two integers, two strings, or two complex numbers.

Operator overloading is a powerful feature that can make your code more concise and readable. It can also make your code more flexible, as it allows you to use the same operator with different data types without having to write separate functions for each type.

Here are some of the benefits of operator overloading:

Code conciseness: Operator overloading allows you to write more compact code by using the same operator for different operations.
Code readability: Operator overloading can make your code more readable by making it clear what the code is doing.
Code flexibility: Operator overloading can make your code more flexible by allowing you to use the same operator with different data types.
However, there are also some potential drawbacks to operator overloading:

Code complexity: Operator overloading can make your code more complex if you are not careful.
Code confusion: Operator overloading can lead to confusion if you are not careful to use operators consistently.
In general, you should only overload operators if you are sure that it will improve your code. If you are not sure, it is probably best to avoid operator overloading.

Here are some examples of how operator overloading is used in C++:

Adding two integers:int a = 5; int b = 3; int c = a + b;
Concatenating two strings:std::string str1 = "Hello"; std::string str2 = "World"; std::string str3 = str1 + str2;
Multiplying two complex numbers:ComplexNumber c1(1, 2); ComplexNumber c2(3, 4); ComplexNumber c3 = c1 * c2;
As you can see, operator overloading can be used to perform a wide variety of operations on different data types. This makes it a powerful tool for writing concise, readable, and flexible code.
*******************************************************************
o this point in the course, you have only read information from the keyboard and written to the screen. But it is not always that case that your programs will need to input/output data in this way. Some programs require no interaction with humans. It is sometimes the case that a much more convenient way to import data to a program is a text file containing that data. Perhaps data is manipulated and written to a storage file, then read in again to be modified a second time. So it is clear that we need a way to establish this capability. Before showing how it is done, I will explain more precisely just what is happening with these statements:

cin>>var;
cout<<var;
Streams
We discussed operators in the 3rd module. An operator is a function that is represented by a symbol (+, -, &&, !, etc). Like a function, an operator returns a value and has parameters. We call the arguments that are passed to the parameters operands – the operator operates on operands. And like functions, operators can be overloaded. Of course, we are not allowed to redefine the functionality of operators for built-in types. But we can redefine them for types that we invent. This is called operator overloading, and we will soon be addressing that issue.

The << and >> symbols in the statements above are operators. << is the “insertion operator” and >> is the “extraction operator”. And like any other operator, they take operands. They are both binary operators, taking two operands. For the insertion operator, the operands are cout and var; for the extraction operator, the operands are cin and var. For the statement on the left, we say that we are “extracting data from the cin stream, streaming it into the memory for var. For the right hand statement, we say that we are “inserting data into the cout stream from the memory for var”. << and >> are streaming operators, and cin and cout are stream objects. That’s right, cin and cout are variables, objects. They are objects of type istream and ostream, respectively. It is perhaps difficult to think of them in this way, and you don’t have to, but it is an accurate view. These two streams, cin and cout, and the functionality of the insertion and extraction operators, are provided for us automatically by the compiler when we include iostream in our programs. They give us the ability to stream data from the keyboard and to the monitor.

Stream Declarations and Connecting to It
Now that you have a more accurate understanding of streams, we can address streaming from a file. In order to stream data to or from a file, we will have to create a stream object of the right type. First, you must #include <fstream> in the file where you are going to create the file stream. Of course, fstream is short for file stream. There are two types of file streams: ifstream is the type for input file streaming; ofstream is the type for output file streaming. Here is the code needed to create a file stream:

#include <fstream>
ifstream in;    // in is a stream that will stream data from a file
ofstream out;   // out is a stream that will stream data to a file
Notice that these are declarations of variables just like any other declaration: <type> <name>. And now that we have declared the streams, we need to connect them to a file. Let’s concentrate on the ifstream object first.

We will have to assume the existence of a data file in order for this to make sense. You cannot stream data from a file that does not exist. You can create a data file just like you create any other file. The difference is that a data file is pure text containing data only. You certainly do not want to put anything else in a data file, such as a comment header. For the purposes of this demonstration, let’s assume the file is named “input.dat”. Typically the extension .dat is used for data files. And let’s assume it contains nothing but numbers, perhaps like this: