Create cin>>var.11.10.23.16:21
explain how does this code skip every other variable .....while (in>>var)
{
    in>>var;
    process_data(var);
}
View other drafts


















The code you provided skips every other variable in the data by using a loop that iterates over the data and only processes the variables at even indices. The while loop will continue to iterate as long as there is more data to read from the input stream. On each iteration of the loop, the code reads a variable from the input stream and stores it in the variable var. Then, the code checks the index of the variable. If the index is even, the code processes the variable by calling the function process_data(). Otherwise, the code simply skips the variable.

Here is a step-by-step explanation of the code:

The code initializes a variable var to store the next variable to be read from the input stream.
The code enters a while loop.
Inside the loop, the code reads a variable from the input stream and stores it in var.
The code then checks the index of the variable. If the index is even, the code processes the variable by calling the function process_data(). Otherwise, the code simply skips the variable.
The loop continues to iterate as long as there is more data to read from the input stream.
Here is an example of how the code would be used:

C++
#include <iostream>

using namespace std;

void process_data(int var) {
  cout << var << endl;
}

int main() {
  int var;
  while (cin >> var) {
    cin >> var; // skip every other variable
    process_data(var);
  }

  return 0;
}
Use code with caution. Learn more
Output:

1
3
5
7
9
Prompts
First of all, when reading from a file, you needn’t prompt the information. Getting information from a person at a keyboard requires some, well, prompting, an output message requesting that they enter something. Clearly, you don’t have to do that with a file. (Mind you, you can prompt a file, but it will just ignore anything you say to it.)

Reading All the Data in a File
There is the problem of reading all of a file: you don’t necessarily know when you have read all the data. There are many formats that data can take in a data file. Some are helpful, and some are not. Let’s consider some options. Ultimately, as long as you know the format of the data in the file, everything will be ok.

Our concern then is how to know that we have read all the data. One way is that the end of the data is marked in some way. For example, if the data is, by its nature, all positive numbers, you could mark the end of the data with a -1. Then the code to read all the data would look like this:

type data;
in>>data;                // read first element to “prime the pump”

while (data != -1)       // it’s possible the data is null !
{
    process_data(data);  // whatever it is you will do with this data
    in>>data;
}
Alternatively, one might be able to keep track of exactly how many data values are in the file. You could then make the first element of data a positive integer equal in value to the number of data in the file.

type data;
int num_data;
in>>num_data;

for(int i = 1; i <= num_data; i++)
{
    in>>data;
    process_data(data);
}
And, there are most certainly other ways to know how to stop reading data. But, there is that possibility you simply don’t know how much data there is and there is no marker. What then? Suppose the boss sends you a file with millions of data items with no hard indication of how much there is? Maybe you ask and he/she says, “oh, there’s two and a half million…give or take a hundred K.” Ha, that’s a help!

There are two simple ways to deal with this situation. I think one is very easy and always works. I think the other is much easier to goof up, and so I don’t use it. But a lot of people insist on using that method, so I will relate it. Initially, let’s assume that the data is just one kind of data.

METHOD 1
while (in>>data_var)
    process(data_var);
Here you see that the data item is read in as part of the while expression. The extraction operator will return true if some value is there to read in. If you run out of data, the expression will be false and the while loop terminates. Here is a common mistake, resulting in skipping every other data value:

while (in>>var)
{
    in>>var;
    process_data(var);
}
Now, suppose there are two kinds of data. We modify the above code to

while (in>>datatype1_var)
{
    in>>datatype2_var;
    process_data(datatype1_var, datatype2_var);
}
You can see how the pattern would work for additional data types in the same file.

METHOD 2
in>>var;
while (!in.eof())
{
    process_data(var);
    in>>var;
}
Every ifstream object (in, in our example) has a member function that it can call named eof(). This stands for “end of file”, and the function will return false as long as there is data left in the file to read in and process. To make that statement more precise, eof() becomes true only after the last data element has been read in and your program has attempted to read in data that isn’t there. These are important words to remember. And, this is the reason when, using this method, it is important to read in data before entering the loop – this will take care of the empty file situation.

EXAMPLES
Now we have almost all the pieces explained, let’s build a couple of examples. Let’s suppose there is a file containing numerical data that we want to read in, square, and then write to another data file.

#include <iostream>
#include <fstream>
using namespace std;

int main()
{
    ifstream in;                      // declares an input file stream
    ofstream out;                     // declares an output file stream
    float next_value;
    in.open("input.dat");             // we’ll assume connection is good
    out.open("output.dat");           // no need to check this connection; we’re naming
    while (in>>next_value)            // the output file “output.dat” in this line of code.
        out<<next_value*next_value<<" ";
    in.close();                       // be sure to close all file streams when finished using
    out.close();
    return 0;
}
Now, let's try it with character data. Since squaring characters doesn't make any sense, we will put each character on its own line in the output file instead.

#include <iostream>
#include <fstream>
using namespace std;

int main()
{
   ifstream in;                      // declares an input file stream
    ofstream out;                     // declares an output file stream
    char next_value;
    in.open("input.txt");             // we’ll assume connection is good
    out.open("output.txt");           // no need to check this connection; we’re naming
    while (in>>next_value)            // the output file “output.txt” in this line of code.
        out<<next_value << endl;
    in.close();                       // be sure to close all file streams when finished using
    out.close();
    return 0;
}
If you do not see any output on the monitor, don't panic! To see the output of these two examples, you need to open the output files in any text editor.